add Main and SalesManages

+++ src/SalesManages.java
@@ -1,13 +1,13 @@
 public class SalesManages {
-    protected long[] sales;
+    protected int[] sales;
 
-    public SalesManages(long[] sales) {
+    public SalesManages(int[] sales) {
         this.sales = sales;
     }
 
-    public long max() {
-        long max = -1;
-        for (long sale : sales) {
+    public int max() {
+        int max = -1;
+        for (int sale : sales) {
             if (sale > max) {
                 max = sale;
             }
